Ticket #2040

+++ modules/boonex/timeline/classes/BxTimelineResponse.php
@@ -85,7 +85,8 @@ class BxTimelineResponse extends BxBaseModNotificationsResponse
                 //--- Delete cached
                 $this->_oModule->_oDb->deleteCache(array('context_id' => 0)); //--- Delete cache for Public feed
                 $this->_oModule->_oDb->deleteCache(array('context_id' => $aEvent[$CNF['FIELD_OWNER_ID']])); //--- Delete cache for old context
-                $this->_oModule->_oDb->deleteCache(array('context_id' => $aParamsSet['owner_id'])); //--- Delete cache for new context
+                if(isset($aParamsSet['owner_id']))
+                    $this->_oModule->_oDb->deleteCache(array('context_id' => $aParamsSet['owner_id'])); //--- Delete cache for new context
                 break;
 
             case BX_BASE_MOD_NTFS_HANDLER_TYPE_DELETE:
